改进：1. 创建主机时可以绑定指定eip，或者新建eip并且绑定。2. 取消绑定eip时候可以自动删除eip

+++ pkg/compute/models/storagecaches.go
@@ -488,7 +488,7 @@ func (cache *SStoragecache) SyncCloudImages(
 			syncResult.DeleteError(err)
 		} else {
 			syncResult.Delete()
-			if image.getStoragecacheCount() == 0 {
+			if image != nil && image.getStoragecacheCount() == 0 {
 				err = image.Delete(ctx, userCred)
 				if err != nil {
 					log.Errorf("image delete error %s", err)
